Changed code formatting and warning message
