+++ resources/lang/en-US.yml
@@ -1836,4 +1836,4 @@ msg_town_plots_type_line: '<dark_green>%s: <green>%s <gray>/ <aqua>%s <gray>/ <y
 msg_town_plots_type_line_revenue: ' <gray>/ <dark_green>%s'
 
 #Message giving confirmation warning for trusttown
-confirmation_msg_trusttown_consequences: 'Are you sure you wish to trust all residents in this town? This may include future residents of this town if not revoked.'
+confirmation_msg_trusttown_consequences: 'You are about to give full plot permissions to an entire town. This means that the mayor of the town will be able to decide who gets permission to build and break in your town, by adding residents. If this town is open, then anyone can join and gain full access to your town. Are you sure you wish to trust all residents (including future residents,) from this town?'
